1.0-MySQL/requirements.txt
fastapi[standard]>=0.115.13
uvicorn[standard]>=0.34.3
SQLAlchemy>=2.0.41
pydantic>=2.11.7
pydantic-settings>=2.0.0
python-dotenv>=0.19.0
PyMySQL>=1.0.0
python-multipart>=0.0.5 